requirements.txt
attrs==22.1.0
Babel==2.12.1
bcrypt==4.0.1
blinker==1.5
boto3==1.25.2
botocore==1.28.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.0
Flask==2.2.2
flask-babel==3.1.0
Flask-Breadcrumbs==0.5.1
Flask-Dance==7.0.0
Flask-Dropzone==1.6.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Menu==0.7.2
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
pdfkit==1.0.0
pluggy==1.0.0
pymongo==4.3.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
python-slugify==8.0.1
pytz==2023.3
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
sendgrid==6.9.7
six==1.16.0
slugify==0.0.1
starkbank-ecdsa==2.2.0
stripe==4.2.0
text-unidecode==1.3
tomli==2.0.1
urllib3==1.26.12
URLObject==2.4.3
waitress==2.0.0
Werkzeug==2.2.2
wkhtmltopdf==0.2
WTForms==3.0.1